requirements.txt
appdirs==1.4.3
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.6
docutils==0.14
google-api-python-client==1.6.4
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
olefile==0.44
packaging==16.8
Pillow==4.3.0
psycopg2==2.7.3.1
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycrypto==2.6.1
pyparsing==2.2.0
pytz==2017.2
requests==2.18.4
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
